tests: Ensure integrity of the database
Our data source format is nice to work with in the library (build the
database, use a simple ORM, etc...).

However, it we're human, and we might end up making mistakes while
maintaining this data set.

Consider the following lines:

  《              0       0       0       0       0       0       0	1       0       H       yyyag   "       6
  《              0       0       0       0       0       0       0	1       0       H       zxbu    "       6

They represent the same Chinese character, which can have two different
Cangjie codes.

However, a lot of information is duplicated between these two lines: the
Simplified Chinese equivalent, the orientation, and all the booleans.

These are attributes of this Chinese character, and all occurences of
this character in the database MUST have the same values for these
fields.

This test ensures that when we modify one of these attributes for a
certain line in a certain file, we won't forget to change it everywhere.