context: Add properties and constructor
We now have a proper constructor, which sets the appropriate GObject
properties for the CangjieContext object.

Those properties can also be changed at run-time. I'd like to see if we
can get that to work nicely, so that API consumers don't have to always
destroy/recreate a new CangjieContext when the user changes the settings
or something.