context: Check orientation value
This is a bit weird, so it deserves some explanation... (which at some
point should certainly go into the API documentation, when we have one)

We have characters in a database. Some of the can be used only to input
vertical text, others only for horizontal texts.

But most of them can be used to input BOTH vertically and horizontally.
(this is the case of every single Chinese character, in fact).

This explains the 3 possible values for the CangjieOrientation enum.

However, a user is inputting either vertically, or horizontally. It
makes no sense whatsoever to say that a user is inputting both
vertically and horizontally at the same time.

As such, only 2 values are correct for the CangjieContext.

When the user inputs horizontally, we will return all the characters
which are either horizontal or "both vertical and horizontal". (i.e
those NOT vertical)

And when the user inputs vertically, we will return all the characters
which are either vertical or "both vertical and horizontal". (i.e those
NOT horizontal)