README: Remove the license badge
This badge only works with Github, so it broke now that we moved to
Gitlab.

In any case, the licensing of this project is detailed in its own
section of the README.

Finally, Gitlab already shows the license in the project metadata at the
top of the project page with a link to the license file, which serves
the purpose of this badge perfectly.